expose intercept tab as an opt-in devtools plugin
