+++ .changeset/intercept-plugin.md
@@ -1,0 +1,5 @@
+---
+"@osdk/react-devtools": patch
+---
+
+expose the intercept tab as an opt-in devtools plugin via the advanced entry point
